# flights/internal/dto/flights.go
package dto

import (
	"time"

	"github.com/google/uuid"
)

type GetFlightsRequestQuery struct {
	FlightNumber *string `validate:"omitempty,max=8"`
	Origin       *string `validate:"omitempty,max=3,alpha"`
	Destination  *string `validate:"omitempty,max=3,alpha"`
	Status       *string `validate:"omitempty,oneof=Scheduled CheckIn Boarding Delayed Departed Arrived Cancelled Redirected"`
	DateFrom     *time.Time
	DateTo       *time.Time

	Limit int `validate:"omitempty,min=1,max=100"`
	Page  int `validate:"omitempty,min=1"`
}

type CreateFlightRequest struct {
	FlightNumber string    `json:"flight_number" validate:"required,max=8"`
	Origin       string    `json:"origin" validate:"required,len=3,alpha"`
	Destination  string    `json:"destination" validate:"required,len=3,alpha"`
	Date         time.Time `json:"date" validate:"required,gt=now"`
	Status       string    `json:"status,omitempty" validate:"omitempty,oneof=Scheduled CheckIn Boarding Delayed Departed Arrived Cancelled Redirected"`
	Aircraft     string    `json:"aircraft" validate:"required,gt=0"`
}

type PatchFlightRequest struct {
	ID           uuid.UUID  `json:"id" validate:"required"`
	FlightNumber *string    `json:"flight_number,omitempty" validate:"omitempty,max=8"`
	Origin       *string    `json:"origin,omitempty" validate:"omitempty,len=3,alpha"`
	Destination  *string    `json:"destination,omitempty" validate:"omitempty,len=3,alpha"`
	Date         *time.Time `json:"date,omitempty" validate:"omitempty"`
	Status       *string    `json:"status,omitempty" validate:"omitempty"`
	Aircraft     *string    `json:"aircraft,omitempty" validate:"omitempty,gt=0"`
}

type FlightResponse struct {
	ID           uuid.UUID `json:"id"`
	FlightNumber string    `json:"flight_number"`
	Origin       string    `json:"origin"`
	Destination  string    `json:"destination"`
	Date         time.Time `json:"date"`
	Status       string    `json:"status"`
	Aircraft     string    `json:"aircraft"`
}

type PaginationResponse struct {
	Page  int `json:"page"`
	Limit int `json:"limit"`
}

type FlightListResponse struct {
	Data       []FlightResponse   `json:"data"`
	Pagination PaginationResponse `json:"pagination"`
}